# gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  public.GetAllPublicRes:
    properties:
      publics:
        items:
          $ref: '#/definitions/public.Public'
        type: array
    type: object
...
  public.Public:
    properties:
      birthday:
        type: string
      created_at:
        type: string
      deleted_at:
        type: integer
      email:
        type: string
      gender:
        type: string
      id:
        type: string
      last_name:
        type: string
      name:
        type: string
      party_id:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
  public.PublicCreate:
    properties:
      birthday:
        type: string
      email:
        type: string
      gender:
        type: string
      id:
        type: string
      last_name:
        type: string
      name:
        type: string
      party_id:
        type: string
      phone:
        type: string
    type: object
  public.PublicUpdate:
    properties:
      birthday:
        type: string
      email:
        type: string
      gender:
        type: string
      id:
        type: string
      last_name:
        type: string
      name:
        type: string
      party_id:
        type: string
      phone:
        type: string
    type: object
...
  vote.GetAllPublicVoteRes:
    properties:
      count:
        type: integer
      public_votes:
        items:
          $ref: '#/definitions/vote.PublicVoteRes'
        type: array
    type: object
  vote.GetAllVoteRes:
    properties:
      count:
        type: integer
      votes:
        items:
          $ref: '#/definitions/vote.VoteRes'
        type: array
    type: object
...
  vote.PublicVoteCreate:
    properties:
      candidate_id:
        type: string
      election_id:
        type: string
      id:
        type: string
      public_id:
        type: string
    type: object
  vote.PublicVoteRes:
    properties:
      created_at:
        type: string
      election_id:
        type: string
      id:
        type: string
      public_id:
        type: string
    type: object
  vote.Void:
    type: object
  vote.VoteRes:
    properties:
      created_at:
        type: string
      id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for managing online voting system resources
  title: Online Voting System API
  version: "1.0"
paths:
...
  /public:
    delete:
      consumes:
      - application/json
      description: Deletes a public by its ID.
      parameters:
      - description: Public ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vote.Void'
        "400":
          description: Invalid request parameters
          schema:
            type: string
        "404":
          description: Resource not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete a public
      tags:
      - Publics
    post:
      consumes:
      - application/json
      description: Creates a new public.
      parameters:
      - description: Public data
        in: body
        name: public
        required: true
        schema:
          $ref: '#/definitions/public.PublicCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/public.Public'
        "400":
          description: Invalid request body
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create a new public
      tags:
      - Publics
    put:
      consumes:
      - application/json
      description: Updates an existing public.
      parameters:
      - description: Public data
        in: body
        name: public
        required: true
        schema:
          $ref: '#/definitions/public.PublicUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vote.Void'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Resource not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Update a public
      tags:
      - Publics
  /public/all:
    get:
      consumes:
      - application/json
      description: Retrieves all publics.
      parameters:
      - description: Party ID (optional)
        in: query
        name: party_id
        type: string
      - description: Name (optional)
        in: query
        name: name
        type: string
      - description: Last name (optional)
        in: query
        name: last_name
        type: string
      - description: Phone (optional)
        in: query
        name: phone
        type: string
      - description: Email (optional)
        in: query
        name: email
        type: string
      - description: Birthday (optional)
        in: query
        name: birthday
        type: string
      - description: Gender (optional)
        in: query
        name: gender
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/public.GetAllPublicRes'
        "400":
          description: Invalid request parameters
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get all publics
      tags:
      - Publics
  /public/id:
    get:
      consumes:
      - application/json
      description: Retrieves a public by its ID.
      parameters:
      - description: Public ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/public.Public'
        "400":
          description: Invalid request parameters
          schema:
            type: string
        "404":
          description: Resource not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get a public by its ID
      tags:
      - Publics
  /public_vote:
    post:
      consumes:
      - application/json
      description: Creates a new public vote.
      parameters:
      - description: Public vote data
        in: body
        name: public_vote
        required: true
        schema:
          $ref: '#/definitions/vote.PublicVoteCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vote.PublicVoteRes'
        "400":
          description: Invalid request body
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create a new public vote
      tags:
      - Public Votes
  /public_vote/public/all:
    get:
      consumes:
      - application/json
      description: Retrieves all public votes.
      parameters:
      - description: Election ID (optional)
        in: query
        name: election_id
        type: string
      - description: Public ID (optional)
        in: query
        name: public_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vote.GetAllPublicVoteRes'
        "400":
          description: Invalid request parameters
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get all public votes
      tags:
      - Public Votes
  /public_vote/public/id:
    get:
      consumes:
      - application/json
      description: Retrieves a public vote by its ID.
      parameters:
      - description: Public vote ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vote.PublicVoteRes'
        "400":
          description: Invalid request parameters
          schema:
            type: string
        "404":
          description: Resource not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get a public vote by public vote ID
      tags:
      - Public Votes
  /public_vote/vote/all:
    get:
      consumes:
      - application/json
      description: Retrieves all votes.
      parameters:
      - description: Candidate ID (optional)
        in: query
        name: candidate_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vote.GetAllVoteRes'
        "400":
          description: Invalid request parameters
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get all votes
      tags:
      - Public Votes
  /public_vote/vote/id:
    get:
      consumes:
      - application/json
      description: Retrieves a public vote by its ID.
      parameters:
      - description: Vote ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vote.VoteRes'
        "400":
          description: Invalid request parameters
          schema:
            type: string
        "404":
          description: Resource not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get a public vote by vote ID
      tags:
      - Public Votes
swagger: "2.0"